+++ store/src/main/java/discord4j/store/service/StoreService.java
@@ -38,11 +38,11 @@ public interface StoreService {
     /**
      * Returns an arbitrary priority for this service. This is used for automated service discovery in the case
      * that multiple services are present. Conventions: 0 = neutral priority,
-     * {@link Short#MAX_VALUE} = lowest priority, {@link Short#MIN_VALUE}.
+     * {@link Integer#MAX_VALUE} = lowest priority, {@link Integer#MIN_VALUE}.
      *
      * @return The priority of this service, 0 by default.
      */
-    default short priority() {
+    default int priority() {
         return 0;
     }
